Treat the "mper" attribute as the 64-bit item that it is rather than a 32-bit item as hithereto. Output it as a hexadecimal number to correspond with the format of the track id obtained from AppleScript

+++ metadata_hub.h
@@ -68,7 +68,7 @@ typedef struct metadata_bundle {
   char *cover_art_pathname;
   int cover_art_pathname_changed;
 
-  uint32_t item_id; // seems to be a track ID -- see itemid in DACP.c
+  uint64_t item_id; // seems to be a track ID -- see itemid in DACP.c
   int item_id_changed;
   int item_id_received; // important for deciding if the track information should be ignored.
